fix(safety,beat,ui): AUDIT_2 gated behavior fixes -- SOS honest path, ghost win-condition, delta math, FAB overlap (F-1, F-3, F-4, U-1, U-4)
F-1 (Critical) -- SOS previously showed "Sending SOS" / "Alerting
contacts" and then, if the SMS composer failed to launch, silently
returned to the map with zero signal to the user. On-device testing
confirmed this exact failure. SafetyService.sendSos now returns a
real SosResult (opened/failedToOpen) instead of swallowing the
outcome, builds one composer with all contacts as recipients instead
of firing N intents in a loop, and the SOS dialog:
  - relabels the countdown honestly ("Prepare emergency message" /
    "Opening your messaging app...") -- nothing here ever auto-sends
  - on success, confirms the composer opened and still needs a tap
    to send
  - on failure, shows a real failure dialog with a one-tap "Call
    now" fallback to the first emergency contact (SafetyService.
    callContact), which itself falls back to a visible error if even
    the dialer can't open
Verified end-to-end on-device: both the messaging-app failure and
the follow-on dialer failure now surface honestly instead of the
previous silent no-op.

F-3 (High) -- ghost-race "wins" only ever compared average pace, with
no check that the ghost's actual distance was covered -- stopping a
marathon race after 150m at a fast pace registered as a full win.
live_dashboard.dart's finish handler now requires covering >=97% of
the ghost's distance before pace is even considered; a run that
stops short is threaded through as `didNotFinish` to
GhostResultScreen, which shows a distinct "Race incomplete" state
(neither win nor loss framing, no "Try a harder tier" CTA, honest
share text) instead of misreporting it as a loss.

F-4 (High) -- ghostExpectedTimeAtDistance's split interpolation
summed take(lowerIndex+1)/take(upperIndex+1) as its lower/upper
bounds, both of which already included time from the split the
runner is still inside of, overstating elapsed time by roughly one
split for most of a race (DISCOVERED_ISSUES.md #5). Fixed to
take(lowerIndex)/take(lowerIndex+1); the two "KNOWN BUG" locking
tests now assert the corrected values.

U-1 (Medium) -- the idle-state Play/Start button was an absolutely-
positioned overlay with a fixed bottom offset sized for a much
shorter selector sheet than the current 2x2 activity grid, so it
visually landed on top of the Walk/Cycle row. Moved it into the
normal Column flow directly after ActivityTypeSelector so it can't
overlap the grid regardless of sheet height.

U-4 (Low) -- added a touch more top padding under the Learn page's
app bar; the hero title read as flush/cramped against the bar.

flutter analyze clean; 67/67 app tests pass (66 existing + new
did-not-finish regression test).

Ref: AUDIT_2.md.

+++ app/lib/features/learn/learn_page.dart
@@ -25,7 +25,9 @@ class LearnPage extends ConsumerWidget {
       body: CustomScrollView(
         slivers: [
           SliverPadding(
-            padding: const EdgeInsets.fromLTRB(AppTheme.s24, AppTheme.s8, AppTheme.s24, AppTheme.s24),
+            // U-4: a touch more breathing room between the app bar and the
+            // hero title -- s8 read as flush/cramped against the bar.
+            padding: const EdgeInsets.fromLTRB(AppTheme.s24, AppTheme.s16, AppTheme.s24, AppTheme.s24),
             sliver: SliverList(
               delegate: SliverChildListDelegate([
                 Text(L10n.tr('academy', locale), style: text.headlineMedium),
